requirements.txt
requests==2.32.3
python-dotenv==1.0.1
python-telegram-bot==22.7
streamlit==1.45.0
folium==0.19.3
streamlit-folium==0.22.2
